App.jsxの整理
