+++ src/styles/NavigationHeaderStyles.jsx
@@ -1,0 +1,15 @@
+const commonHeaderOptions = {
+  headerStyle: {
+    backgroundColor: '#C0C0C0',
+    height: 120,
+  },
+  headerTitle: 'Men Men',
+  headerTintColor: '#000000',
+  headerBackTitle: 'Back',
+  headerTitleStyle: {
+    fontStyle: 'italic',
+    fontSize: 35,
+  },
+};
+
+export default commonHeaderOptions;
